Move MediaScannerServiceUtils to service.scanner package
